feat(deps): update dependencies and deprecated fields (#279)
* feat(deps): update dependencies and deprecated fields

This commit updates several dependencies to their latest versions, including Kotlin, Gradle, and the Places SDK. It also replaces the deprecated `Place.Field.NAME` with `Place.Field.DISPLAY_NAME` throughout the test suite.

* feat: Update target SDK to 36

This commit updates the `targetSdk` and `compileSdk` to version 36. It also adds the `POST_NOTIFICATIONS` permission to the demo app's `AndroidManifest.xml`, due to the Glide dependency.

* fix: back to SDK version 34

* fix(build): update test workflow

Updates the test workflow to use newer versions of the following actions:
- actions/checkout@v4
- gradle/actions/wrapper-validation@v4
- actions/setup-java@v4.2.1

* fix(build): revert coreKtx to 1.13.1

* chore: added hilt compiler to catalog

---------

Co-authored-by: Enrique López Mañas <eenriquelopez@gmail.com>